refactor(cli): extract sandbox skill install action
